+++ sites/ademe/CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Configure Marsha video as LTI provider instead of "jisc.ac.uk" test platform
 - Rename `LTI_TEST_*` settings to `LTI_*` as "TEST" does not make sense here
 
 ## [0.1.0] - 2021-08-16